requirements.txt
fastapi
alembic
sqlalchemy
pydantic
loguru
python-dotenv
asyncpg
psycopg2
uvicorn
passlib
bcrypt
pyjwt
python-multipart
aiohttp